skills/chandra-ocr/tests/test_extract.py
#!/usr/bin/env python3
"""
chandra-ocr — 单元测试

用法:
    pytest tests/test_extract.py -v
"""

import pytest
import tempfile
import os
from pathlib import Path

try:
    from chandra import extract
except ImportError:
    pytest.skip("未安装 chandra-ocr", allow_module_level=True)


class TestExtract:
    """测试文档提取功能"""
    
    def test_extract_pdf_to_markdown(self):
        """测试 PDF 转 Markdown"""
        # 注意：需要真实的 PDF 测试文件
        # 这里使用示例代码结构
        pytest.skip("需要真实 PDF 测试文件")
    
    def test_extract_pdf_to_html(self):
        """测试 PDF 转 HTML"""
        pytest.skip("需要真实 PDF 测试文件")
    
    def test_extract_pdf_to_json(self):
        """测试 PDF 转 JSON"""
        pytest.skip("需要真实 PDF 测试文件")
    
    def test_extract_image(self):
        """测试图片提取"""
        pytest.skip("需要真实图片测试文件")


class TestBatchProcess:
    """测试批量处理功能"""
    
    def test_batch_process_directory(self):
        """测试批量处理目录"""
        pytest.skip("需要真实测试文件")
    
    def test_batch_process_concurrent(self):
        """测试并发处理"""
        pytest.skip("需要真实测试文件")


class TestFormats:
    """测试输出格式"""
    
    def test_markdown_format(self):
        """测试 Markdown 格式输出"""
        pytest.skip("需要真实测试文件")
    
    def test_html_format(self):
        """测试 HTML 格式输出"""
        pytest.skip("需要真实测试文件")
    
    def test_json_format(self):
        """测试 JSON 格式输出"""
        pytest.skip("需要真实测试文件")


class TestLanguages:
    """测试多语言支持"""
    
    def test_chinese_simplified(self):
        """测试简体中文识别"""
        pytest.skip("需要中文测试文件")
    
    def test_english(self):
        """测试英文识别"""
        pytest.skip("需要英文测试文件")
    
    def test_japanese(self):
        """测试日文识别"""
        pytest.skip("需要日文测试文件")


class TestSpecialContent:
    """测试特殊内容识别"""
    
    def test_table_recognition(self):
        """测试表格识别"""
        pytest.skip("需要含表格的测试文件")
    
    def test_form_recognition(self):
        """测试表单识别"""
        pytest.skip("需要含表单的测试文件")
    
    def test_handwriting_recognition(self):
        """测试手写识别"""
        pytest.skip("需要手写测试文件")
    
    def test_math_formula(self):
        """测试数学公式识别"""
        pytest.skip("需要含公式的测试文件")


if __name__ == "__main__":
    pytest.main([__file__, "-v"])
